refactor(packages): init GeoJSON adapter package
