+++ lab/package.json
@@ -10,7 +10,6 @@
     "test": "ava -v"
   },
   "dependencies": {
-    "fuse.js": "^3.4.5",
-    "ramda": "^0.26.1"
+    "@cadaster/adapter-geojson-fulltext": "*"
   }
 }
