+++ packages/adapter-geojson-fulltext/index.js
